updated lexer and tokens

+++ fiberc/src/token/literal.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone, PartialEq)]
 pub enum Literal {
-    Integer(u32),
+    Integer(u64),
     Float(f32),
     Boolean(bool),
     Character(char),
